% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pteridocat}
\alias{pteridocat}
\title{Taxonomic database of Pteridophytes}
\format{
Tibble, including the following columns:
\itemize{
\item \code{taxonID}: Unique identifier for the set of taxon information.
\item \code{acceptedNameUsageID }: Identifier for the name usage (documented meaning
of the name according to a source) of the currently accepted taxon; maps to
\code{taxonID}.
\item \code{taxonomicStatus}: The status of the use of the \code{scientificName} as a label
for a taxon.
\item \code{scientificName}: The full scientific name, with authorship information.
\item \code{namePublishedIn}: A reference for the publication in which the
\code{scientificName} was originally established under the rules of the
International Code of Nomenclature for algae, fungi, and plants
\item \code{taxonRemarks}: Notes about the taxon.
}
}
\source{
\url{https://dwc.tdwg.org/terms/#taxon}
}
\usage{
pteridocat
}
\description{
Taxonomic database of pteridophytes following
\href{https://dwc.tdwg.org/terms/#taxon}{Darwin Core standards for taxonomic data}.
}
\details{
Description of data below modified from
\href{https://www.tdwg.org/}{Biodiversity Information Standards (TDWG)} under
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International License}.
}
\keyword{datasets}
