% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pteridocat}
\alias{pteridocat}
\title{Taxonomic database of Pteridophytes}
\format{
Dataframe (tibble), including the following columns:
\itemize{
\item \code{taxonID}: Unique identifier for the set of taxon information.
\item \code{parentNameUsageID}: Identifier for the name usage (documented meaning of
the name according to a source) of the direct, most proximate higher-rank
parent taxon (in a classification) of the most specific element of the
\code{scientificName} (i.e., parent taxon); maps to \code{taxonID}.
\item \code{acceptedNameUsageID }: Identifier for the name usage of the currently
accepted taxon (i.e., accepted name of a synonym); maps to
\code{taxonID}.
\item \code{taxonomicStatus}: The status of the use of the \code{scientificName} as a label
for a taxon.
\item \code{taxonRank}: The taxonomic rank of the most specific name in the
\code{scientificName.}
\item \code{scientificName}: The full scientific name, with authorship information.
\item \code{genericName}: The genus part of the \code{scientificName} without authorship.
\item \code{infragenericEpithet}: The infrageneric part of a binomial name at ranks
above species but below genus.
\item \code{specificEpithet}: The name of the first or species epithet of the
\code{scientificName.}
\item \code{infraspecificEpithet}: The name of the lowest or terminal infraspecific
epithet of the scientificName, excluding any rank designation.
\item \code{namePublishedIn}: A reference for the publication in which the
\code{scientificName} was originally established under the rules of the
\code{nomenclaturalCode}.
\item \code{nomenclaturalCode}: The nomenclatural code under which the
\code{scientificName} is constructed; typically, "ICN", for
International Code of Nomenclature for Algae, Fungi, and Plants.
\item \code{nomenclaturalStatus}: The status related to the original publication of
the name and its conformance to the relevant rules of nomenclature.
\item \code{taxonRemarks}: Notes about the taxon.
\item \code{references}: 	A related resource that is referenced, cited, or otherwise
pointed to by the described resource.
\item \code{modified}: The most recent date-time on which the information was changed.
(NA for data that were imported from Catalog of Life and not modified).
\item \code{nameAccordingTo}: The reference to the source in which the specific taxon
concept circumscription is defined or implied.
}
}
\source{
Hassler, M. (2021). Checklist of Ferns and Lycophytes of the World.
In O. Bánki, Y. Roskov, L. Vandepitte, R. E. DeWalt, D. Remsen, P. Schalk,
T. Orrell, M. Keping, J. Miller, R. Aalbu, R. Adlard, E. Adriaenssens, C.
Aedo, E. Aescht, N. Akkari, M. A. Alonso-Zarazaga, B. Alvarez, F. Alvarez,
G. Anderson, et al., Catalogue of Life Checklist (Version 2021-08-06).
\url{https://doi.org/10.48580/d4sd-3dc}
}
\usage{
pteridocat
}
\description{
Taxonomic database of pteridophytes following
\href{https://dwc.tdwg.org/terms/#taxon}{Darwin Core standards for taxonomic data}.
}
\details{
Taxonomic data of pteridophytes compiled by Hassler (2021) were downloaded
from the \href{https://www.catalogueoflife.org/}{Catalog of Life}
and modified under the
\href{http://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution (CC BY) 4.0 license}.

Description of data format modified from
\href{https://www.tdwg.org/}{Biodiversity Information Standards (TDWG)} under the
\href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution (CC BY) 4.0 license}.
}
\keyword{datasets}
